update coords to be floats

+++ src/libcolonyfind/colony_finder.py
@@ -366,8 +366,8 @@ def generate_baseplate_coords(coords, cam_x = CONSTANTS.CAM_X, cam_y = CONSTANTS
                     # colony_coord[0] = ((colony_coord[0] - center_x)/img_width) * cam_x # FIXME THIS IS PROBABLY WRONG
                     # colony_coord[1] = ((colony_coord[1] - center_y)/img_height) * cam_y
 
-                    colony_coord[0] = colony_coord[0] / img_width
-                    colony_coord[1] = colony_coord[1] / img_height
+                    colony_coord[0] = float(colony_coord[0] / img_width)
+                    colony_coord[1] = float(colony_coord[1] / img_height)
 
 
                     colony_coord = colony_coord[:-1] # remove radius from colony coord
